// packages/automl/frontend/src/app/types/pipeline.ts
export enum RuntimeStateKF {
  RUNTIME_STATE_UNSPECIFIED = 'RUNTIME_STATE_UNSPECIFIED',
  PENDING = 'PENDING',
  RUNNING = 'RUNNING',
  SUCCEEDED = 'SUCCEEDED',
  SKIPPED = 'SKIPPED',
  FAILED = 'FAILED',
  CANCELING = 'CANCELING',
  CANCELED = 'CANCELED',
  CACHED = 'CACHED',
  PAUSED = 'PAUSED',
}
// ...
export type InputOutputDefinition = {
  artifacts?: InputOutputDefinitionArtifacts;
  parameters?: Record<string, unknown>;
};
// ...
export type DAG = {
  tasks: Record<string, TaskKF>;
  outputs?: unknown;
};

export type PipelineComponentKF = {
  dag?: DAG;
  executorLabel?: string;
  inputDefinitions?: InputOutputDefinition;
  outputDefinitions?: InputOutputDefinition;
};

export type PipelineComponentsKF = Record<string, PipelineComponentKF | undefined>;

export type PipelineExecutorKF = {
  container?: {
    image: string;
    command?: string[];
    args?: string[];
  };
  [key: string]: unknown;
};

export type PipelineExecutorsKF = Record<string, PipelineExecutorKF | undefined>;

export type PipelineSpec = {
  components: PipelineComponentsKF;
  deploymentSpec: { executors: PipelineExecutorsKF };
  pipelineInfo: { name: string; description?: string };
  root: {
    dag: DAG;
    inputDefinitions?: InputOutputDefinition;
    outputDefinitions?: InputOutputDefinition;
  };
  schemaVersion: string;
  sdkVersion: string;
};

export type PlatformSpec = {
  platforms: {
    kubernetes?: {
      deploymentSpec?: { executors?: PipelineExecutorsKF };
    };
  };
};

export type PipelineSpecVariable = {
  pipeline_spec?: PipelineSpec;
  platform_spec?: PlatformSpec;
} & Partial<PipelineSpec>;